tests/requirements.txt
aiogram~=3.14.0
aiogram-cli~=1.1.0
aiohttp~=3.10.10
uvicorn~=0.32.0
requests~=2.32.3
pytest-benchmark~=5.1.0
